Merges runners together s.t. there's only one
- Removes generic sleeps, now polls each test server to see if ready
